+++ process_responses/process_output.py
@@ -14,20 +14,18 @@ class ResponseProcessor:
     '''
     Process remote server responses to server, ledger, and validation subscription stream messages.
 
-    :param settings: Config file
-    :param asyncio.queues.Queue message_queue: Incoming websocket messages
-    :param asyncio.queues.Queue notification_queue: Outbound SMS messages
+    :param dict args_d: Default settings, tables, and queues
     '''
-    def __init__(self, settings, table_stock, table_validator, message_queue, notification_queue):
-        self.settings = settings
-        self.table_stock = table_stock
-        self.table_validator = table_validator
+    def __init__(self, args_d):
+        self.settings = args_d['settings']
+        self.table_stock = args_d['table_stock']
+        self.table_validator = args_d['table_validator']
         self.forks = []
         self.ll_modes = []
         self.val_keys = []
         self.processed_validations = []
-        self.message_queue = message_queue
-        self.notification_queue = notification_queue
+        self.message_queue = args_d['message_queue']
+        self.notification_queue = args_d['notification_queue']
         self.time_last_output = 0
         self.time_fork_check = 0
         self.last_heartbeat = time.time()
@@ -144,3 +142,25 @@ class ResponseProcessor:
                 break
             except Exception as error:
                 logging.critical(f"Otherwise uncaught exception in response processor: '{error}'.")
+
+def start_output_processing(args_d):
+    '''
+    Start the asyncio loop.
+
+    :param dict args_d: Default settings, queues, and tables.
+    '''
+    loop = asyncio.new_event_loop()
+    monitor_tasks = []
+
+    try:
+        monitor_tasks.append(
+            loop.create_task(ResponseProcessor(args_d).process_messages())
+        )
+
+        logging.warning("Response processor loop started.")
+        loop.run_forever()
+
+    except KeyboardInterrupt:
+        for task in monitor_tasks:
+            task.cancel()
+        logging.critical("Closed response processor asyncio loops.")
