+++ process_responses/process_stock_output.py
@@ -113,16 +113,6 @@ async def check_state_change(server, message, notification_queue):
             }
         )
 
-async def log_keys(message_result, table):
-    '''
-    Keep track of all the potential keys returned by server subscription messages.
-    '''
-    logging.warning("Checking for new keys.")
-    ignore_keys = ['type',]
-    for key in message_result.keys():
-        if key not in table[0].keys() and key not in ignore_keys:
-            logging.warning(f"new server subscription key found: '{key}'.")
-
 async def update_table_server(table, notification_queue, message):
     '''
     Add info contained in new messages to the table.
@@ -137,8 +127,6 @@ async def update_table_server(table, notification_queue, message):
         message_result = message['data']['result']
     elif message['data'].get('type') == 'serverStatus':
         message_result = message['data']
-
-    await log_keys(message_result, table)
 
     for server in table:
         if server['url'] == message['server_url']:
